Move handlers & routes for results receiver
into the receiver package for consistency.
